+++ bin/lib/test/test.utils.js
@@ -1,7 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-var test = require("tape");
-test("Gazeall", function (t) {
-    t.equal("Watching gazeall run!", "Watching gazeall run!");
-    t.end();
-});